Added login UI with logout fix

+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { Nav, Platform,NavController } from 'ionic-angular';
+import { Nav, Platform } from 'ionic-angular';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
 
@@ -24,6 +24,7 @@ export class MyApp {
     this.pages = [
       { title: 'Home', component: HomePage },
       { title: 'List', component: ListPage },
+      { title: 'Logout', component: null}
     ];
 
   }
@@ -37,14 +38,19 @@ export class MyApp {
     });
   }
 
-  logout(){
-    localStorage.clear();
-    this.nav.popToRoot()
-  }  
+  // logout(){
+  //   localStorage.clear();
+  //   this.nav.setRoot(LoginPage);
+  // }  
 
   openPage(page) {
     // Reset the content nav to have just this page
     // we wouldn't want the back button to show in this scenario
-    this.nav.setRoot(page.component);
+    if(page.component) {
+        this.nav.setRoot(page.component);
+    } else {
+        localStorage.clear();
+        this.nav.setRoot(LoginPage);
+    }
   }
 }
